+++ routes/_app.tsx
@@ -10,11 +10,10 @@ export default function App({ Component }: AppProps) {
           rel='stylesheet'
           href='https://cdn.jsdelivr.net/gh/CarcajadaArtificial/lunchbox@0.1.1/static/style.css'
         />
+        <link rel='stylesheet' href='style.css' />
       </Head>
       <body class='clr-bg-panel clr-txt-base txt-paragraph'>
-        <div class='_screen'>
-          <Component />
-        </div>
+        <Component />
       </body>
     </>
   );