+++ routes/index.tsx
@@ -5,7 +5,6 @@ import { InputNote } from '../islands/InputNote/index.tsx';
 export default function Home() {
   return (
     <Main layout_type='center'>
-      <Text type='heading'>Input Note</Text>
       <InputNote />
     </Main>
   );